+++ examples/mobile-client/voip-call/app/src/hooks/useRecentsRedial.ts
@@ -1,0 +1,25 @@
+import { useVoip } from '@fishjam-cloud/react-native-client';
+import { useEffect } from 'react';
+
+import { useUser } from '../user';
+import { usePlaceCall } from './usePlaceCall';
+
+/**
+ * Places a call when the user redials from the iOS Recents list. The SDK holds the
+ * intent until we are ready, so we can simply wait for the session to be restored.
+ */
+export function useRecentsRedial(): void {
+  const { pendingCallIntent, clearCallIntent } = useVoip();
+  const { username } = useUser();
+  const placeCall = usePlaceCall();
+
+  useEffect(() => {
+    if (!pendingCallIntent || !username) return;
+
+    const { handle } = pendingCallIntent;
+    clearCallIntent();
+    placeCall(handle).catch((err) =>
+      console.error('[voip] failed to start call from a Recents intent:', err),
+    );
+  }, [pendingCallIntent, username, clearCallIntent, placeCall]);
+}
